requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
albumentations=1.2.0=pypi_0
appnope=0.1.4=pypi_0
asttokens=3.0.0=pypi_0
backcall=0.2.0=pyhd3eb1b0_0
ca-certificates=2025.2.25=hecd8cb5_0
certifi=2025.1.31=pypi_0
charset-normalizer=3.4.1=pypi_0
comm=0.2.1=py38hecd8cb5_0
contourpy=1.1.1=pypi_0
cycler=0.12.1=pypi_0
debugpy=1.6.7=py38hcec6c5f_0
decorator=5.2.1=pypi_0
executing=2.2.0=pypi_0
filelock=3.16.1=pypi_0
fonttools=4.56.0=pypi_0
fsspec=2025.2.0=pypi_0
huggingface-hub=0.29.2=pypi_0
idna=3.10=pypi_0
imageio=2.35.1=pypi_0
importlib-metadata=7.0.1=py38hecd8cb5_0
importlib-resources=6.4.5=pypi_0
importlib_metadata=7.0.1=hd3eb1b0_0
ipdb=0.13.13=pypi_0
ipykernel=6.29.5=py38hecd8cb5_0
ipython=8.12.3=pypi_0
jedi=0.19.2=pypi_0
jinja2=3.1.6=pypi_0
joblib=1.4.2=pypi_0
jupyter_client=8.6.0=py38hecd8cb5_0
jupyter_core=5.7.2=py38hecd8cb5_0
kiwisolver=1.4.7=pypi_0
libcxx=14.0.6=h9765a3e_0
libffi=3.4.4=hecd8cb5_1
libsodium=1.0.18=h1de35cc_0
markupsafe=2.1.5=pypi_0
matplotlib=3.7.5=pypi_0
matplotlib-inline=0.1.7=pypi_0
mpmath=1.3.0=pypi_0
ncurses=6.4=hcec6c5f_0
nest-asyncio=1.6.0=py38hecd8cb5_0
networkx=3.1=pypi_0
numpy=1.24.4=pypi_0
opencv-python=4.11.0.86=pypi_0
opencv-python-headless=4.11.0.86=pypi_0
openssl=3.0.16=h184c1cd_0
packaging=24.2=pypi_0
parso=0.8.4=pypi_0
pexpect=4.9.0=pypi_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=10.4.0=pypi_0
pip=25.0.1=pypi_0
platformdirs=3.10.0=py38hecd8cb5_0
prompt-toolkit=3.0.50=pypi_0
psutil=5.9.0=py38hca72f7f_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure-eval=0.2.3=pypi_0
pure_eval=0.2.2=pyhd3eb1b0_0
pygments=2.19.1=pypi_0
pyparsing=3.1.4=pypi_0
python=3.8.20=hce00570_0
python-dateutil=2.9.0post0=py38hecd8cb5_2
pywavelets=1.4.1=pypi_0
pyyaml=6.0.2=pypi_0
pyzmq=25.1.2=py38hcec6c5f_0
qudida=0.0.4=pypi_0
readline=8.2=hca72f7f_0
regex=2024.11.6=pypi_0
requests=2.32.3=pypi_0
safetensors=0.5.3=pypi_0
scikit-image=0.18.3=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.10.1=pypi_0
setuptools=75.3.0=pypi_0
six=1.17.0=pypi_0
sqlite=3.45.3=h6c40b1e_0
stack-data=0.6.3=pypi_0
stack_data=0.2.0=pyhd3eb1b0_0
sympy=1.13.3=pypi_0
threadpoolctl=3.5.0=pypi_0
tifffile=2023.7.10=pypi_0
tk=8.6.14=h4d00af3_0
tokenizers=0.20.3=pypi_0
tomli=2.2.1=pypi_0
torch=2.2.2=pypi_0
torchvision=0.17.2=pypi_0
tornado=6.4.1=py38h46256e1_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=py38hecd8cb5_0
transformers=4.46.3=pypi_0
typing-extensions=4.12.2=pypi_0
typing_extensions=4.11.0=py38hecd8cb5_0
urllib3=2.2.3=pypi_0
wcwidth=0.2.13=pypi_0
wheel=0.44.0=py38hecd8cb5_0
xz=5.6.4=h46256e1_1
zeromq=4.3.5=hcec6c5f_0
zipp=3.20.2=py38hecd8cb5_0
zlib=1.2.13=h4b97444_1
